refactor xcm files + fix tests
