+++ src/pages/accountForm/AccountForm.test.tsx
@@ -91,6 +91,9 @@ describe("AccountForm", () => {
           selectedAccount: {},
         },
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
   });
