+++ src/pages/balance/components/Activity.test.tsx
@@ -36,6 +36,9 @@ describe("Actvity", () => {
           selectedAccount: selectedEVMAccountMock,
         },
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
 
     vi.mock("@src/Extension");
