+++ src/pages/balance/components/ChainSelector.test.tsx
@@ -56,6 +56,9 @@ describe("ChainSelector", () => {
         },
         setSelectNetwork: () => setSelectNetwork(),
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
     vi.mock("react-router-dom", () => ({
       useNavigate: () => vi.fn(),
