+++ src/pages/manageAssets/ManageAssets.test.tsx
@@ -28,6 +28,9 @@ describe("ManageAssets", () => {
           type: AccountType.EVM,
         },
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
 
     vi.mock("react-router-dom", () => ({
