+++ src/pages/send/components/ConfirmTx.test.tsx
@@ -32,6 +32,9 @@ describe("ConfirmTx", () => {
           selectedAccount: selectedEVMAccountMock,
         },
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
 
     vi.mock("react-router-dom", () => ({
