+++ src/pages/send/components/EvmForm.test.tsx
@@ -50,6 +50,9 @@ describe("EvmForm", () => {
           ],
         },
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
 
     vi.mock("react-hook-form", () => ({
