+++ src/pages/settings/Security.test.tsx
@@ -22,6 +22,9 @@ describe("Security", () => {
       useAccountContext: () => ({
         getSelectedAccount: () => getSelectedAccount(),
       }),
+      useThemeContext: () => ({
+        color: "red",
+      }),
     }));
 
     vi.mock("react-router-dom", () => ({
