+++ src/xcm/assets.ts
@@ -1,0 +1,21 @@
+import { PARACHAINS } from "@src/constants/chains";
+import {
+  MOONBEAM_ASSETS_MAPPING,
+  ASTAR_ASSETS_MAPPING,
+  ACALA_ASSETS_MAPPING,
+  SHIDEN_ASSETS_MAPPING,
+} from "./chains";
+
+interface IXCM_ASSETS_MAPPING {
+  [key: string]: {
+    [key: string]: string[];
+  };
+}
+
+export const XCM_ASSETS_MAPPING: IXCM_ASSETS_MAPPING = {
+  [PARACHAINS.MOONBEAM]: MOONBEAM_ASSETS_MAPPING,
+  [PARACHAINS.ASTAR]: ASTAR_ASSETS_MAPPING,
+  [PARACHAINS.ACALA]: ACALA_ASSETS_MAPPING,
+  [PARACHAINS.MOONRIVER]: MOONBEAM_ASSETS_MAPPING,
+  [PARACHAINS.SHIDEN]: SHIDEN_ASSETS_MAPPING,
+};
