+++ src/xcm/chains/index.ts
@@ -1,0 +1,6 @@
+export * from "./kusama";
+export * from "./polkadot";
+export * from "./moonbeam";
+export * from "./acala";
+export * from "./astar";
+export * from "./shiden";
