+++ src/xcm/extrinsics.ts
@@ -1,0 +1,26 @@
+import { PARACHAINS, RELAY_CHAINS } from "@src/constants/chains";
+import { IXCM_MAPPING } from "./interfaces";
+import {
+  POLKADOT_EXTRINSICS,
+  KUSAMA_EXTRINSICS,
+  MOONBEAM_EXTRINSICS,
+  ACALA_EXTRINSICS,
+  ASTAR_EXTRINSICS,
+  SHIDEN_EXTRINSICS,
+} from "./chains";
+
+export const XCM_MAPPING: IXCM_MAPPING = {
+  [RELAY_CHAINS.POLKADOT]: POLKADOT_EXTRINSICS,
+
+  [RELAY_CHAINS.KUSAMA]: KUSAMA_EXTRINSICS,
+
+  [PARACHAINS.MOONBEAM]: MOONBEAM_EXTRINSICS,
+
+  [PARACHAINS.ACALA]: ACALA_EXTRINSICS,
+
+  [PARACHAINS.ASTAR]: ASTAR_EXTRINSICS,
+
+  [PARACHAINS.MOONRIVER]: MOONBEAM_EXTRINSICS,
+
+  [PARACHAINS.SHIDEN]: SHIDEN_EXTRINSICS,
+};
