+++ vite.config.ts
@@ -47,6 +47,7 @@ export default defineConfig({
         "**/src/constants/**",
         "**/src/routes/**",
         "**/src/components/common/**",
+        "**/src/xcm/**",
       ],
     },
   },
